Add files via upload
Since we did this project ages ago, it is a bit of a mess. However, this still might be useful to some people. Do give it a glance, and setting it up properly could be a hassle so I can be contacted in case of any issues. I couldn't upload some videos (we used them in the background) due to the large size. At the same time, you need to install some packages:

Run this:
npm init -y

and
npm install express